improve validator integration and fix linter
Wanted a solution that
- lintable
- uses the generated code
- keeps the type safety
- the related parts are colocated
- works like normal ts code when used

To achieve this:
- removed the generated code from eslint
- moved all the schema type definitions into the validate folder
- moved the validator function export and typing here
- barrelled all the validators into a single module export
