+++ .eslintrc.js
@@ -67,7 +67,7 @@ module.exports = {
       },
     ],
   },
-  ignorePatterns: ["dist/", "extension-reload.js"],
+  ignorePatterns: ["dist/", "extension-reload.js", "**/validate/*.js"],
   parser: "@typescript-eslint/parser",
   parserOptions: {
     project: "./.tsconfig-eslint.json",
