+++ background/lib/alchemy.ts
@@ -14,44 +14,11 @@ import {
 import { ETH } from "../constants"
 import { getEthereumNetwork } from "./utils"
 import { AnyEVMTransaction, EVMNetwork } from "../networks"
-import { JTDDataType, ValidateFunction } from "ajv/dist/jtd"
-
-// JSON Type Definition for the Alchemy assetTransfers API.
-// https://docs.alchemy.com/alchemy/documentation/enhanced-apis/transfers-api
-//
-// See RFC 8927 or jsontypedef.com to learn more about JTD.
-const alchemyAssetTransferJTD = {
-  properties: {
-    asset: { type: "string", nullable: true },
-    hash: { type: "string" },
-    blockNum: { type: "string" },
-    category: { enum: ["token", "internal", "external"] },
-    from: { type: "string", nullable: true },
-    to: { type: "string", nullable: true },
-    erc721TokenId: { type: "string", nullable: true },
-  },
-  optionalProperties: {
-    rawContract: {
-      properties: {
-        address: { type: "string", nullable: true },
-        decimal: { type: "string", nullable: true },
-        value: { type: "string", nullable: true },
-      },
-    },
-  },
-  additionalProperties: true,
-} as const
-
-const alchemyGetAssetTransfersJTD = {
-  properties: {
-    transfers: {
-      elements: alchemyAssetTransferJTD,
-    },
-  },
-  additionalProperties: true,
-} as const
-
-const isValidAlchemyAssetTransferResponse: ValidateFunction<JTDDataType<typeof alchemyGetAssetTransfersJTD>> = require("./validate/jtd-validators.js")["alchemy-get-asset-transfers.jtd.schema.json"]
+import {
+  isValidAlchemyAssetTransferResponse,
+  isValidAlchemyTokenBalanceResponse,
+  isValidAlchemyTokenMetadataResponse,
+} from "./validate"
 
 /**
  * Use Alchemy's getAssetTransfers call to get historical transfers for an
@@ -148,28 +115,6 @@ export async function getAssetTransfers(
     .filter((t): t is AssetTransfer => t !== null)
 }
 
-// JSON Type Definition for the Alchemy token balance API.
-// https://docs.alchemy.com/alchemy/documentation/enhanced-apis/token-api
-//
-// See RFC 8927 or jsontypedef.com for more detail to learn more about JTD.
-const alchemyTokenBalanceJTD = {
-  properties: {
-    address: { type: "string" },
-    tokenBalances: {
-      elements: {
-        properties: {
-          contractAddress: { type: "string" },
-          tokenBalance: { type: "string", nullable: true },
-          error: {},
-        },
-      },
-    },
-  },
-  additionalProperties: false,
-} as const
-
-const isValidAlchemyTokenBalanceResponse: ValidateFunction<JTDDataType<typeof alchemyTokenBalanceJTD>> = require("./validate/jtd-validators.js")["alchemy-token-balance.jtd.schema.json"]
-
 /**
  * Use Alchemy's getTokenBalances call to get balances for a particular address.
  *
@@ -229,22 +174,6 @@ export async function getTokenBalances(
       })
   )
 }
-
-// JSON Type Definition for the Alchemy token metadata API.
-// https://docs.alchemy.com/alchemy/documentation/enhanced-apis/token-api#alchemy_gettokenmetadata
-//
-// See RFC 8927 or jsontypedef.com for more detail to learn more about JTD.
-const alchemyTokenMetadataJTD = {
-  properties: {
-    decimals: { type: "uint32" },
-    name: { type: "string" },
-    symbol: { type: "string" },
-    logo: { type: "string", nullable: true },
-  },
-  additionalProperties: false,
-} as const
-
-const isValidAlchemyTokenMetadataResponse : ValidateFunction<JTDDataType<typeof alchemyTokenMetadataJTD>> = require("./validate/jtd-validators.js")["alchemy-token-metadata.jtd.schema.json"]
 
 /**
  * Use Alchemy's getTokenMetadata call to get metadata for a token contract on
