+++ background/lib/prices.ts
@@ -1,5 +1,4 @@
 import { fetchJson } from "@ethersproject/web"
-import { JSONSchemaType, ValidateFunction } from "ajv"
 import logger from "./logger"
 import {
   AnyAsset,
@@ -10,46 +9,9 @@ import {
 } from "../assets"
 
 import { toFixedPoint } from "./fixed-point"
+import { isValidCoinGeckoPriceResponse } from "./validate"
 
 const COINGECKO_API_ROOT = "https://api.coingecko.com/api/v3"
-
-export type CoingeckoPriceData = {
-  [coinId: string]:
-    | {
-        last_updated_at: number
-        [currencyId: string]: number | undefined
-      }
-    | undefined
-}
-
-// Ajv's typing incorrectly requires nullable: true for last_updated_at because
-// the remaining keys in the coin entry are optional. This in turn interferes
-// with the fact that last_updated_at is listed in `required`. The two `as`
-// type casts below trick the type system into allowing the schema correctly.
-// Note that the schema will validate as required, and the casts allow it to
-// match the corret TypeScript types.
-//
-// This all stems from Ajv also incorrectly requiring an optional property (`|
-// undefined`) to be nullable (`| null`). See
-// https://github.com/ajv-validator/ajv/issues/1664, which should be fixed in
-// Ajv v9 via
-// https://github.com/ajv-validator/ajv/commit/b4b806fd03a9906e9126ad86cef233fa405c9a3e
-export const coingeckoPriceSchema: JSONSchemaType<CoingeckoPriceData> = {
-  type: "object",
-  required: [],
-  additionalProperties: {
-    type: "object",
-    properties: {
-      last_updated_at: { type: "number" } as {
-        type: "number"
-        nullable: true
-      },
-    },
-    required: ["last_updated_at"] as never[],
-    additionalProperties: { type: "number", nullable: true },
-    nullable: true,
-  },
-}
 
 export async function getPrice(
   coingeckoCoinId = "ethereum",
@@ -58,13 +20,12 @@ export async function getPrice(
   const url = `${COINGECKO_API_ROOT}/simple/price?ids=${coingeckoCoinId}&vs_currencies=${currencySymbol}&include_last_updated_at=true`
 
   const json = await fetchJson(url)
-  const validate: ValidateFunction<JSONSchemaType<typeof coingeckoPriceSchema>> = require("./validate/json-validators.js")
 
-  if (!validate(json)) {
+  if (!isValidCoinGeckoPriceResponse(json)) {
     logger.warn(
       "CoinGecko price response didn't validate, did the API change?",
       json,
-      validate.errors
+      isValidCoinGeckoPriceResponse.errors
     )
 
     return null
@@ -88,13 +49,12 @@ export async function getPrices(
   const json = await fetchJson(url)
   // TODO fix loss of precision from json
   // TODO: TESTME
-  const validate: ValidateFunction<JSONSchemaType<typeof coingeckoPriceSchema>> = require("./validate/json-validators.js")
 
-  if (!validate(json)) {
+  if (!isValidCoinGeckoPriceResponse(json)) {
     logger.warn(
       "CoinGecko price response didn't validate, did the API change?",
       json,
-      validate.errors
+      isValidCoinGeckoPriceResponse.errors
     )
 
     return []
