+++ background/redux-slices/0x-swap.ts
@@ -1,11 +1,16 @@
 import { createSlice, createSelector } from "@reduxjs/toolkit"
 import { fetchJson } from "@ethersproject/web"
 import { utils } from "ethers"
+import { JTDDataType, ValidateFunction } from "ajv/dist/jtd"
 
 import { createBackgroundAsyncThunk } from "./utils"
 import { Asset, FungibleAsset, isSmartContractFungibleAsset } from "../assets"
 import logger from "../lib/logger"
-import { JTDDataType, ValidateFunction } from "ajv/dist/jtd"
+import {
+  isValidSwapAssetsResponse,
+  isValidSwapPriceResponse,
+  isValidSwapQuoteResponse,
+} from "../lib/validate"
 
 interface PartialSwapAssets {
   sellAsset?: FungibleAsset
@@ -91,23 +96,6 @@ export const initialState: SwapState = {
   zrxPrices: [],
 }
 
-const swapAssetsJTD = {
-  properties: {
-    records: {
-      elements: {
-        properties: {
-          address: { type: "string" },
-          decimals: { type: "int8" },
-          name: { type: "string" },
-          symbol: { type: "string" },
-        },
-      },
-    },
-  },
-}
-
-const isValidSwapAssetsResponse: ValidateFunction<JTDDataType<typeof swapAssetsJTD>> = require("../lib/validate/jtd-validators.js")["swap-assets.jtd.schema.json"]
-
 export const fetchSwapAssets = createBackgroundAsyncThunk(
   "0x-swap/fetchAssets",
   async () => {
@@ -126,24 +114,6 @@ export const fetchSwapAssets = createBackgroundAsyncThunk(
     return []
   }
 )
-
-const swapPriceJTD = {
-  properties: {
-    records: {
-      elements: {
-        properties: {
-          price: { type: "string" },
-          symbol: { type: "string" },
-        },
-      },
-    },
-    page: { type: "uint32" },
-    perPage: { type: "uint32" },
-    total: { type: "uint32" },
-  },
-}
-
-const isValidSwapPriceResponse: ValidateFunction<JTDDataType<typeof swapPriceJTD>> = require("../lib/validate/jtd-validators.js")["swap-price.jtd.schema.json"]
 
 export const fetchSwapPrices = createBackgroundAsyncThunk(
   "0x-swap/fetchPrices",
@@ -166,52 +136,6 @@ export const fetchSwapPrices = createBackgroundAsyncThunk(
   }
 )
 
-const swapQuoteJTD = {
-  properties: {
-    allowanceTarget: { type: "string" },
-    buyAmount: { type: "string" },
-    buyTokenAddress: { type: "string" },
-    buyTokenToEthRate: { type: "string" },
-    chainId: { type: "uint32" },
-    data: { type: "string" },
-    estimatedGas: { type: "string" },
-    gas: { type: "string" },
-    gasPrice: { type: "string" },
-    guaranteedPrice: { type: "string" },
-    minimumProtocolFee: { type: "string" },
-    orders: {
-      elements: {
-        properties: {
-          makerAmount: { type: "string" },
-          makerToken: { type: "string" },
-          source: { type: "string" },
-          sourcePathId: { type: "string" },
-          takerAmount: { type: "string" },
-          takerToken: { type: "string" },
-          type: { type: "uint32" },
-        },
-      },
-    },
-    price: { type: "string" },
-    protocolFee: { type: "string" },
-    sellAmount: { type: "string" },
-    sellTokenAddress: { type: "string" },
-    sellTokenToEthRate: { type: "string" },
-    sources: {
-      elements: {
-        properties: {
-          name: { type: "string" },
-          proportion: { type: "string" },
-        },
-      },
-    },
-    to: { type: "string" },
-    value: { type: "string" },
-  },
-}
-
-const isValidSwapQuoteResponse: ValidateFunction<JTDDataType<typeof swapQuoteJTD>> = require("../lib/validate/jtd-validators.js")["swap-quote.jtd.schema.json"]
-
 export const fetchSwapQuote = createBackgroundAsyncThunk(
   "0x-swap/fetchQuote",
   async (quote: { assets: SwapAssets; amount: SwapAmount }) => {
@@ -226,8 +150,6 @@ export const fetchSwapQuote = createBackgroundAsyncThunk(
         `buyToken=${quote.assets.buyAsset.symbol}&` +
         `sellAmount=${sellAmount}`
     )
-
-    return apiData as ZrxQuote
 
     if (isValidSwapQuoteResponse(apiData)) {
       return apiData as ZrxQuote
